Removed for loop float increment

+++ RTSClone/RTSClone/PathFinding.cpp
@@ -418,7 +418,7 @@ glm::vec3 PathFinding::getAvailablePositionOutsideAABB(const Entity& senderEntit
 		glm::vec2 position = Globals::convertToGridPosition(senderEntity.getPosition());
 		for (int ray = 1; ray <= Globals::NODE_SIZE * 5; ++ray)
 		{
-			position = position + direction * static_cast<float>(1);
+			position = position + direction * 1.0f;
 			if (!map.isPositionOccupied(position) && map.isWithinBounds(position))
 			{
 				return Globals::convertToWorldPosition(position);
@@ -444,7 +444,7 @@ glm::vec3 PathFinding::getClosestPositionOutsideAABB(const glm::vec3& entityPosi
 	}
 
 	glm::vec3 position = centrePositionAABB;
-	for (float ray = 1.0f; ray <= Globals::NODE_SIZE * 7.0f; ++ray)
+	for (int ray = 1; ray <= Globals::NODE_SIZE * 7; ++ray)
 	{
 		position = position + direction * 1.0f;
 		if (!AABB.contains(position) && !map.isPositionOccupied(position) && map.isWithinBounds(position))
